support big responses for Apache 2 adaptor

+++ Utilities/Adaptors/Adaptor/request.c
@@ -168,7 +168,7 @@ const char *req_AddHeader_common(HTTPRequest *req,const char *key,const char *va
     */
    if ((req->content_length == 0) &&
        ((strcasecmp(key, CONTENT_LENGTH) == 0) || (strcasecmp(key, "content_length") == 0)))	/* get content-length */
-       req->content_length = atoi(value);
+       req->content_length = atol(value);
 
 #ifdef DEBUG
        WOLog(WO_DBG,"(req-hdr) %s: %s",key, value);
@@ -263,8 +263,8 @@ int req_sendRequest(HTTPRequest *req, net_fd socket)
    /* Note that we reuse buffers, and the existing content-data buffer. */
    if (req->content_length > req->content_buffer_size)
    {
-      int total_sent = req->content_buffer_size;
-      int len_read, amount_to_read;
+      long total_sent = req->content_buffer_size;
+      long len_read, amount_to_read;
       req->haveReadStreamedData = 1;
       while (total_sent < req->content_length)
       {
@@ -362,8 +362,8 @@ int req_sendRequest(HTTPRequest *req, net_fd socket)
    /* Note that we reuse buffers, and the existing content-data buffer. */
    if (req->content_length > req->content_buffer_size)
    {
-      int total_sent = req->content_buffer_size;
-      int len_read, amount_to_read;
+      long total_sent = req->content_buffer_size;
+      long len_read, amount_to_read;
       req->haveReadStreamedData = 1;
       while (total_sent < req->content_length && result == 0)
       {
